renames module typing => _typing
