+++ src/dimcat/_typing.py
